requirements.txt
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0